bulk commit

+++ controllers/company.controller.js
@@ -1,4 +1,5 @@
 const _ = require('lodash');
+const Joi = require('joi');
 const CompanyModel = require('../models/company.model');
 
 const getAllCompanies = async (req, res) => {
@@ -8,8 +9,8 @@ const getAllCompanies = async (req, res) => {
 		const companies = await CompanyModel.find()
 			.sort({ createdAt: -1 })
 			.skip((page - 1) * limit)
-			.limit(limit)
-			.populate("jobs");
+			.limit(limit);
+		// .populate("jobs");
 
 		const totalNum = await CompanyModel.countDocuments();
 
@@ -42,23 +43,21 @@ const getCompany = async (req, res) => {
 
 const createCompany = async (req, res) => {
 	try {
-		// const companyObj = _.pick(req.body, ["location", "company_name", "email", "website","tel_Number", "job_listings"]);
-		const companyObj = req.body;
-		// const logo = req.files;
-		// const desc = req.files.description.name;
-		// const logo = req.files["logo"][0];
-		// const desc = req.files["description"][0];
-		// return res.status(201).json(companyObj);
-		// console.log(logo, desc);
+		const companyObj = _.pick(req.body, ["location", "companyLogo", "description", "companyName", "companyType", "email", "website", "telNumber",]);
 
-		// companyObj.logo = logo.filename + path.extname(logo.originalname);
-		// companyObj.description = desc.filename + path.extname(desc.originalname);
+		if (companyObj.jobs && !companyObj.jobs.length)
+			companyObj.jobs = [];
+
+		const { error } = joiValidateCompany(companyObj);
+		if (error)
+			return res.status(400).json(error)
 
 		const company = await CompanyModel.create(companyObj);
 		return res.status(201).json(company);
+
 	} catch (error) {
 		console.error(error);
-		return res.status(500).json({ message: error });
+		return res.status(500).json(error);
 	}
 };
 
@@ -66,11 +65,12 @@ const updateCompany = async (req, res) => {
 	try {
 		const { id } = req.params;
 		const companyObj = req.body;
-		const company = await CompanyModel.findById(id);
-		if (!company)
+
+		const foundCompany = await CompanyModel.findById(id);
+		if (!foundCompany)
 			return res.status(400).json({ message: "Companies does not found" });
 
-		Object.assign(company, companyObj);
+		Object.assign(foundCompany, companyObj);
 		const updated = await company.save();
 		return res.status(201).json(updated);
 	} catch (error) {
@@ -82,6 +82,10 @@ const updateCompany = async (req, res) => {
 const deleteCompany = async (req, res) => {
 	try {
 		const { id } = req.params;
+		const foundCompany = await CompanyModel.findById(id);
+		if (!foundCompany)
+			return res.status(400).json({ message: "Companies does not found" });
+
 		const deletedCompany = await CompanyModel.findByIdAndDelete(id);
 		return res.status(200).json(deletedCompany);
 	} catch (error) {
@@ -96,16 +100,39 @@ const addJobs = async (job) => {
 		const company = await CompanyModel.findById(id);
 		if (!company)
 			return false;
-		const jobs = company.jobs;
-		jobs.push(job);
-		company.jobs = jobs;
+		company.jobs.push(job);
 		await company.save();
 		return true;
 	} catch (error) {
 		console.error(error);
-		return res.status(500).json(error);
+		return false;
+		// return error;
 	}
 };
+
+const joiValidateCompany = (compObj) => {
+	const CompanySchema = Joi.object({
+		companyLogo: Joi.string().required(),
+		description: Joi.string().required(),
+		companyName: Joi.string().required(),
+		companyType: Joi.string().required(),
+		website: Joi.string().required(),
+		location: Joi.object({
+			city: Joi.string().required(),
+			country: Joi.string().required(),
+			state: Joi.string().optional(),
+			zipCode: Joi.string().optional(),
+		}).optional(),
+		telNumber: Joi.object({
+			line: Joi.string().required(),
+			mobile: Joi.string().optional(),
+		}).optional(),
+		email: Joi.string().required().email(),
+		jobs: Joi.array().optional(),
+	});
+
+	return CompanySchema.validate(compObj);
+}
 
 
 module.exports = {
